+++ src/lib/components/Header/Nav/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./Nav.svelte";
+export type { NavItemType, NavLinkType, NavMenuType } from "./types";
